requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.7
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.1.3
django-advanced-filters==1.3.0
django-braces==1.14.0
django-cors-headers==3.3.0
django-crispy-forms==1.7.2
django-elasticsearch-dsl==7.2.1
django-extensions==3.1.5
django-filter==1.1.0
django-haystack==3.1.1
django-jsonfield==1.4.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
duplicity==0.8.12.0
elasticsearch==7.15.2
elasticsearch-dsl==7.4.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
itypes==1.2.0
Jinja2==3.0.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.4.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2018.4
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
systemd-python==234
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
Whoosh==2.7.4
xkit==0.0.0
zipp==1.0.0
